requirements.txt
praw
openai
python-dotenv
reportlab
